refactor: refactor stuff in tagset.c
